feat: address fixmes in protocol parameter fixmes
I also regenerated fixtures, since many of them were stale after the RewardAccount changes.

Running the regeneration script would've clobbered the handrolled fixtures specifically for this case. I also updated the script to fail if the handrolled fixtures fail, since those exist to test a specific case and a regression should be clear

+++ crates/amaru-ledger/src/governance/ratification/stake_pools.rs
@@ -231,7 +231,7 @@ mod tests {
             option::of(any::<u64>()),
             option::of(any::<u64>()),
             option::of(any::<u64>()),
-            option::of(any::<u64>()),
+            option::of(any::<u16>()),
             option::of(any::<u64>()),
             option::of(any_ex_units()),
             option::of(any::<u64>()),
